Use a different scheme to make join IDs unique (fix for prior commit)
The prior fix used an increasing integer for Join IDs, which meant that it
created a brand new Join object for every single instance where anything joined
to anything else.

This change includes the "reason" for initiating the join as part of the ID.  In
most cases, this "reason" is the name of the property in the first class that started
the creation of the join chain.  It solves the problem where two differently named
properties in the starting class end up joining through the same table (think: differently
named attributes in a hangoff attribute table).  Even though the same attribute table
gets joined in twice when generating the SQL, the join IDs differ because the property
name in the first class is different.