Query efficiency improvement
If an object is already loaded, but another query causes it to be included
in DB results, the object fabricator used to not update all_params_loaded.

The problem is that if that second query is not the same as the first, then
all_params_loaded won't know that the second query has been done.  Meaning
that if you do that second query again, it would hit the DB again.

The fix is to update all_params_loaded both when it is a newly loaded object
and when it's an already-loaded object