DeletedRefs are turned back into objects after rollback
If outside code holds on to an object reference after calling delete(), it
gets turned into a UR::DeletedRef to indicate that it's no longer a valid
object.  If the object becomes valid again, either through rollback or creating
a new object with the same ID, that object reference will now be re-used so
others still holding on to the reference won'y have to re-get() the object before
using it.