Workaround for a bug in Perl 5.8 involving custom sorts
If, while in a custom sort block using $a->method and $b->method, and
somewhere deep in the call stack resolving the value for method() it
does another sort using $a->other_method and $b->other_method, that inner
sort can have $a and/or $b set to undef.

The actual crash happened when doing "genome --help".  Seems like
Genome::Command (which isa Command::Tree) within sorted_sub_command_classes
sorts by sub_command_sort_position() on a bunch of class names.  When those
classes eventually generate() (which does a sort on those class metaobjects'
property_name()), it triggers the bug.

I couldn't put together a working testcase in a reasonable length of time, so
you'll just have to trust me on this