Make command subclass resolution faster by being smarter about guessing the right subclass name
When figuring out the right command class to run, it used to first load up all
the modules in the given command class' subdirectory, then pick out the right
one.

Now, it first tries a module with just the params passed in, then tries
substituting the word "Command" in different places in the module name, and
finally does the old behavior of loading up everything in the subdirectory.