Move the _load() method from UR::Value into UR::DataSource::QueryPlan
First step of changing UR::Value to use the Default datasource for loading.

QueryPlan isa UR::Value, but we need to retain the ability to miracle
QueryPlans into existence with _load() early in the query process, rather
than allowing QueryPlans to get loaded through the regualr loading process.
The problem is that in order to use the regular loader, you have to generate
a QueryPlan, which is an issue if you're trying to load a QueryPlan...