De-hackify the initialzation of QueryPlans for Default datasources a bit
For a query on any subclass of UR::Value, set the loading template's final_class_name
to the rule's subject_class_name

This may need to happen for _any_ query against the Default data source, not just
for UR::Value subclasses.  I'll have to add some more tests to see.