Fix the shell exit code when using the --lsf switch
When running under --lsf, the command tries to return a failed exit code of 1
when there are test failures (just as prove does), but the LSF cleanup code
in the END block runs one of more 'bkill' commands, which clears the exit code
back to 0 when the bkill commands run without error.

So, in the END block, save the intended value of $? before cleaning up the lsf
jobs, then restore it after they're cleaned up.