Fix for printing compile error messages when a module fails to autoload
UR::Object::Type::InternalAPI::_load() uses error_message() to complain when
eval "use $modulename"
fails by calling error_message() with $@ as the message.

The error message passed into the sub contains a reference to $@ inside the
arg list @_.  When _get_msgdata does its own eval {}, it clears $@, which
leaves an empty string as the message to be printed in @_.

The fix is to make a copy of the message into a lexical before making calls
to other methods which might wipe out the contents of the message.