Allow bare lists of words as values
This allows someone to do this at the cmdline:
lister-cmd name='some value with spaces',score>2

Before, it was a problem because the shell would gobble the single quotes and
the parser would see the string:
name=some value with spaces,score>2

Which is a syntax error because of the space after the word "some"

This commit splits up the 'value' terminal into a legal first-value
and legal subsequent values.  first-values are what 'value' used to be,
where subsequent values are first-values that aren't keywords.  We
need this restriction to avoid having an ambiguous grammar.

The values in a 'between' expression need to first-values (and not bare
lists of values) for the same reason