error_message, warning_message(), etc always return the last message as a string
This fixes the case of calling error_message with no args as a way of relaying
the last message on some object to another object, to send it to print.
Previously, if called in list context, it would return a list of items like
the message, package, file, and line where the last message was set.  It is
a common idiom from the old messaging API to call a method on an object, and
if there was a problem use copy the error_message from that object into $self
like this:
  $self->error_message($other->error_message());

Also added a new method in UR::ModuleBase called error_message_source() that
returns the data the old method did in list context.