Make non-UR objects valid values in BoolExprs
When a non-UR object is used as a value in constructing a boolexpr, the
property's data_type is converted to a Value-type class before it checks
whether the value isa that Value.

In the case where the property's data_type is 'CODE' (for example), and the
value being passed in is a coderef blessed into some other package, then
the previous code would fail because that value->isa(UR::Value::CODE) was
false.

The fix is to have a method on UR::Object::Property is_valid_storage_for_value()
that first does the simple isa() check.  If the property's data_type is a Value
class, then also check that Value's underlying_data_types()