Added method for resolving values for properties in path spec without values in the rule
In the property replacement step, it adds info about where the inserted globs
(*) are.  _replace_glob_with_values_in_pathname() takes that info, and figures
out what the possible values are using glob() which taking care not to guess at
property values for globs put in the original spec string.