Refactor to move common code out of if/else branches
Both parts did similar things to the data.  Each part of the if/else makes a
subref that applies the transforms for this step and returns data for the
next stage.  This subref accepts a sub that fixes the string offsets in the
.__glob_positions__ list to account for the differences in string length
between the passed-in path spec and the resolved paths