Fix finding the proper table name for a given property name
If we found a property with that name, but the property has no column name,
keep going up the inheritance chain, since this property may be overriding
a parent's initial property definition, which will include a column_name.

This happens with an abstract class that subclassifies by a property.  In
the parent class, the property exists as a "regular" property with a column
name.  In the child classes, the property exists as an overridden property
that is read-only and has no column_name.