When updating a DB row, infer the table structure from the class metadata
Previously, we relied on the metaDB info to construct the SQL for an INSERT.
If there was no metaDB info, then it queried the DB's data dictionary and
defined temporary table/column metadata.

Now it uses the class metadata to infer the column names, as UPDATEs and
DELETEs already do.