Test case for committing when the metaDB info is not current
Previously, if the MetaDB had info for a class' table and that info
was out of date, bad things could happen at commit time.

For inserts, it would omit saving columns that had properties but the
info was missing from the metaDB.  For updates, it would die when checking
if the property was getting saved into a BLOB column.

Now it infers the table structure from the class metadata for inserts, and
also behaves properly for updates