Avoid an exception in QueryPlan for certain queries
Fixes the case of filtering by a via/to object accessor property, and the
'to' points to an id_by/id_class_by property.  Previously, QueryPlan
tries to follow the chain of delegation by following 'to' until it finds
one that's not delegated.  In this case, the id_by/id_class_by is delegated
but has no 'to', so it would throw an exception.

In theory, we could continue joining...  The filter's value indicates what
class we should join to next, and the value's ID provides the join filters.
The problem is that the QueryPlan has no access to the values, only the
rule template.

With this fix, it gets the right data, but may have to perform more than one
query against the data source.