Avoid a memory leak when reading a file bails out early
The individual file iterators were keeping a reference to itself.  If the
reading reached EOF, it gave up its reference by reassigning the closed-over
var to the null_sub, and things were cleaned up properly.

But if the reading bailed out early because of it knew the sorting of the file,
then it was not giving up the reference.

The fix was to not worry about reassigning to null_sub.  The multiplexing
iterator (the next one out) splices the individual reader iterators out
of the list when they stop returning data.