Remove initialization of getopt for is_many properties to empty arrayref.
Creating the empty array prevents any default value specified in the command
from being used, since it will always be overwritten with [] (or the supplied
values).  Getopt::Long will create this arrayref for us if a value is supplied
on the command line (even if it is an empty value).