Experiment with a new indirect accessor implementation
The old code has if/else conditions to try and handle different situations.
It doesn't handle the case where the indirect accessor is via a non-standard
reverse-as many-to-many relationship.

This new code just gets the join list and builds a list of closures that can
get from here to there.  It doesn't yet work in the general case (especially
during bootstrapping), so it sets up a special accessor prop_new() for testing