Fix a bug with multi-char record_separators
If the last read of the file did not include all the characters in
the record_separator, then whatever data was in the last column would
include whatever partial separator characters that were there.

While it is possible that actual column data can include the first few
characters of the record_separator, it's probably pretty rare that it
happend only in the very last line in a file.  That's the only case
where it would fill in the wrong data fo an object.