Improve server response when getting foreign key info
The query used by DBD::Oracle to get foreign key info performs poorly on
Oracle 11 because of a hint in the query.  It seems to work better to run
the same query without the hint.  Since DBD::Oracle isn't set up to plug
in different queries, we need to cut-and-paste it's method to remove the
hint.

The original method gets called for Oracle versions before 11