Change the test in line with the new messaging settings behavior
Previously, a child class' settings were inherited from its parent, but only
the first time the value was checked.  After the first check, the parent value
was copied to the child, and changing the value in the parent would not affect
the child.

The new behavior is for the child to always inherit the value from the parent
unless it has set it's own value.  Setting the value to undef goes back to
inheriting the parent's value.