Fix a bug when joining two child classes that don't have tables but their parents do
The join coercion code keeps a list of joined class=>table mappings, but was
not putting in the mapping for the primary class/table the query was for.  In
the case covered by the test, it was picking the wrong class to look up a
column name for - using the source class of the join instead of the source
class of the whole query.