Fix bug with inherited implied property data_type
An object accessor's id_by creates an implied property (the thing it's id_by)
if the class definition doesn't already have that property explicitly
described.  This auto-created implied property is not given a data_type when
the property normalizer creates it.  Later on,
_complete_class_meta_object_definitions() fills in any remaining holes in the
class definition, including providing a data_type for that property.

Originally, it would always pick the data_type from the remote class' ID
property.

This fixes it to use a parent class' property of the same name (inherited)
to get the data type from first, falling back on the old behavior