function names now link back to where they were defined
It does not make a link for functions that were defined before the $^P
manipulation that happens at the top of ClassBrowser.pm.  These are in
the core UR modules that get loaded before ClassBrowser.pm gets control.

We can probably mitigate the problem by making a data cache file for UR
separately