An object's class must be is_uncachable in order to be unload()ed
The check for is_uncachable() is changed so that during the check for
is_uncachable on parent classes, _all_ parent classes must be is_uncachable
before we decide this one is.

This required changing the parentage of FileMux spawned datasources so they
all inherit from a common class that gets set to is_uncachable = true.

The added test verifies that data sources aren't unloaded after a rollback