Sync/commit properly when AutoCommit is on
Base RDBMS datasources have AutoCommit off, but derived classes can turn it on.
If it's on, we need to explicitly start a transaction before altering any data
in case there's a problem in the commit, so we can rollback the partially saved
changes.

As the comment shows, DBI docs say that starting a transaction when AutoCommit
is on will temporarily turn it off until the transaction is over (commit or
rollback)