value_id_to_values_frozen() can correctly thaw data structures containing UR objects
Previously, it would return things blessed as UR objects, but were clones of
the objects in the object cache.  When this new, thawed data went out of
scope, it would incorrectly trigger DESTROY() on those clones, leading to
Badness.

_fixup_ur_objects_from_thawed_data takes in a list of thawed data and
recursively inspects the contents, swapping out cloned UR objects for
the originals in the object cache.

Change-Id: Ib0a2186356e1ccee5c9efd1c7745e561e927eab5