Explicity set the FK actions in the test
Older versions of SQLite (the change happened some time between 3.6.13
and 3.6.22) have a default ON UPDATE and ON DELETE action of RESTRICT,
while 3.6.22 has a default action of NO ACTION for both.  The SQL parser
on 3.6.22 and earlier doesn't accept the string "NO ACITON" there, so
we'll explicity set the action to something we can control.

Change-Id: Iab30df06e2bbc0133a00610238ab08d524f7da53