Fix a bug involving inserting dummy ID data to the alternate DB
If inserting a row depends on inserting other data with a dummy ID, don't
do it.  This can happen if a test is running with dummy-IDs and no-commit on,
commits part-way through the test (which inserts the data but doesn't commit),
then reloads that dummy ID data.

Normally, objects with dummy IDs don't get saved to the database, but if
a row has a foreign key with a dummy ID in the FK column, that _would_ get
inserted.  This violates the FK constraint, but some databases are configured
to enforce FK constraints only at commit time.

These things are always created by the program and not saved in the database
permanently, so it's safe to just refuse to save anything to the alternate DB
that depends on a dummy-IDed object.

Change-Id: Ibabc885a62e54a8316cd5baf275cc0dde1e8c3fe