Add another heuristic to know when it's working with a new FK
If the FKs have names, it's easy to know how to bundle the individual rows
of a multi-column foreign key together, and discriminate between a table with
2 FKs and a single 2-column FK.

But not all FKs have names.  We've been relying on the first column of each
FK to have ORDINAL_POSITION 1.

Came across a case where a table has a pair of 1-column FKs, and they have
ORDINAL_POSITION 1 and 2, but different names.  This change looks at the FK
names as another way to batch up FKs.

Change-Id: Ib6c706f21119ccff417e3e1214312e347427882f