'owner' is no longer tracked in the MetaDB or its objects
If a data source only wants to track a single schema, it should have a value
for its 'owner' property/method.  In this case, class' table names will be
simple table names.

If a data source wants to track all schemas in a database, it should return
a false value from its 'owner' property.  Class' table names will be strings
like schema.table.

The 'owner' column in the metaDB is no longer a primary key in any table
and is no longer getting filled in.  We can drop the columns later when we
have a plan for migrating the MetaDB.