Use the via-via case only when there's a single property linking the bridge to the result
If there's more than 1 ID property linking the bridge to the result, we
can't use this.  The simple thing to do would be to generate a list of property
names and supply listrefs for each property when get get(), but that's wrong.
That query would mean
    where a in <list_a> and b in <list_b>

but we really want a query like
    where (a = a1 and b = b1) or (a = a2 and b = b2)