Bugfix when List::MoreUtils::PP is used
91b_sets_count_with_changes.t was failing only in TravisCI, and only in Perl
5.8.  For some unknown reason, that one version of Perl uses the pure-perl
module while all the others use the XS module.  This test reliably fails when
the env var LIST_MOREUTILS_PP=1

When the pure-perl version of any() is used, it inserts a function call
frame to evaluate the block that isn't there when the XS version is used.
This extra frame sets @_ to the empty list which causes __changes__ to check
all properties on the object, not just the ones the caller is interested in.

The XS version probably uses a shortcut to call the block that doesn't set up
a proper function call with its own @_.