Roles' overloads are composed into classes
Roles with overloading follows the same rules as methods.  If multiple roles
overload the same operator, composing them is disallowed unless the composing
class also overloads the same operator.

The mechanism for discovering what is overloaded and how breaks the
encapsulation of the overload module because it doesn't provide an api for
a caller to discover it.  Overloaded operations are discovered by looking for
subs named like PackageName::(operation.

Overloads are installed into the composing class using the normal
"use overload" mechanism.