Add an attribute handler for RoleParam
This is how parameters for roles will work.  Scalar variables with the
RoleParam attribute will initially be a UR::Role::Param instance.

After getting composed into a role, these variables will become tied and
will resolve to their proper value for the given invocant's class, but that's
not working yet.