Role params are working as 'our' variables
When a role is composed into a role, the property data is cloned rather than
used as-is so the UR::Role::Param instances can be swapped out with the values
provided in the params list each time the role is composed.  After the role
param values are filled in, the original role param variables with the
RoleParam attribute are swapped out with tie() implementations.  They FETCH
their values by searching up the call stack for the first function whose
invocant is a class composing the role the function came from, then retrieving
the param from the UR::Role::Instance created when the role was composed.