Delay calling a calculated default sub until needed
Don't call the sub for a calculated default property until the default value
is actually requested.  This means the calculation is called once immediately
before getting the value for the first time.  It also means the calculation
is never called if a value is set before the value is retrieved.