Properly handle -limit and -offset when there are changed objects
If there are changed objects matching the bx, then we need to do the
full query in the DB (without -limit/-offset).  The explanation appears
in the source, but is reproduced here:

If there are any cached objects, then the asked-for offset may not necessarily
be the offset that applies to data in the database.  And since the offset is not
meaningful, neither is the limit.  Consider a query matching these objects with
limit => 2, offset => 1
In memory: 1 2
In DB    :     3 4 5 6
The result should be (2, 3).  If we kept the -offset in the DB's SQL, we would
have missed object 3.
Similarly, if any DB rows exist as objects with changed data, then rows returnd
from the DB might not be included in the results, and the supplied -limit could
keep us from reading rows that should be returned

Additionally, if -limit/-offset were specified when there are changed objects,
then we can't be sure all the expected results were returned from the DB.
This means we shouldn't force an exception on objects we thought might have been
deleted - maybe we just missed them due to the -limit or -offset.