Fix a bug involving the AutoUnloadPool and iterators
The Context Loading Iterator buffers one object to ensure that all the joined
rows have been loaded before returning the primary object via the iterator.
This interacted badly with the AutoUnloadPool when the pool was created before
the object was pulled off of the iterator by also unloading the buffered object.

Now, the Loading Iterator sets a flag on the buffered object that tells the
AutoUnloadPool to skip unloading that object.  The flag is cleared before the
object gets returned up the stack.