Fix a bug involving the AutoUnloadPool and iterators
The Context Loading Iterator buffers one object to ensure that all the joined
rows have been loaded before returning the primary object via the iterator.
This interacted badly with the AutoUnloadPool when the pool was created before
the object was pulled off of the iterator by also unloading the buffered object.

AutoUnloadPool now weakens references for dropped objects instead of unloading
then explicitly.  Since the Loading Iterator retains a reference to any objects
it buffers, the object doesn't get unloaded until the refcount goes to 0.