Catch the case where a property chain includes a non-existent property
This exception directs the user toward what the problem is.

Previously, the chain resolution code would stop at the first unknown property
and return what it had figured out at that point.  That could lead to an
exception in UR::DS::QueryPlan::_init_core() or _init_rdbms about calling a
method on an undefined value (if the last resolved property was an id-by
property), or resolving to the wrong property (if the last one was a via/to)