Fix a bug with the cross-datasource joiner
Other parts of the system that deal with multiple joins (QueryPlan) have this
same tweak.  The other cross-datasource test doesn't trip this bug because
its classes have ID properties named something other than 'id'.  In this test,
the 'id' property of ThingOne conflicts with the 'id' property of UR::Value
when the ImportIterator is building %foreign_property_name_map.