Ref addresses can be re-used, use a serial number instead
In some cases, the refaddr of the generated iterator code ref can be recycled
between different get() instances.  When that happens, the code thinks it's
within the same get() and neglects to reset the file position with seek().

Some perl builds seem to use smaller address spaces than others, judging on
the size of a refaddr.  Maybe it is just more likely to happen on builds
with smaller address spaces.  For example:
    perl -e 'my $a = \1; print "$a\n"'
    OS X: SCALAR(0x7fd852825dc8)
    Ubuntu: SCALAR(0x27fb958)
Even though they're build 64-bit perls with the same ptrsize.

This fixes #126