Fix test failure on some versions of perl
travis tests are failing for perl <= 5.20.  It was discussed on the
debian-bugs-dist mailing list:
https://www.mail-archive.com/debian-bugs-dist@lists.debian.org/msg1609439.html
and perl5-porters:
https://lists.gt.net/perl/porters/337326

The failure shows as some numbers being quoted in the JSON output and some
numbers are bare.  The test expects quotes.  I was able to reproduce the
failure on 5.20.3, and success on 5.22.4 I'm guessing that some JSON
serializers are more strict about what's a number and what's a string.

The change forces the __display_name__ for UR::Value::Number to be returned
as a numeric value, changes the object's id to be a Number, and changes the
expected test output to have bare numbers in the JSON data.